+++ dwarfeng-dct-core/src/main/java/com/dwarfeng/dct/node/configuration/ValueCodingQosDefinitionParser.java
@@ -1,0 +1,84 @@
+package com.dwarfeng.dct.node.configuration;
+
+import com.dwarfeng.dct.impl.handler.ValueCodingQosHandlerImpl;
+import com.dwarfeng.dct.impl.service.ValueCodingQosServiceImpl;
+import com.dwarfeng.dct.sdk.util.BeanDefinitionParserUtil;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.config.BeanReference;
+import org.springframework.beans.factory.config.ConstructorArgumentValues;
+import org.springframework.beans.factory.config.RuntimeBeanReference;
+import org.springframework.beans.factory.support.AbstractBeanDefinition;
+import org.springframework.beans.factory.support.BeanDefinitionBuilder;
+import org.springframework.beans.factory.support.ManagedMap;
+import org.springframework.beans.factory.xml.BeanDefinitionParser;
+import org.springframework.beans.factory.xml.ParserContext;
+import org.w3c.dom.Element;
+
+import javax.annotation.Nonnull;
+
+/**
+ * 值编码 Qos 元素的 BeanDefinitionParser。
+ *
+ * @author DwArFeng
+ * @since 3.0.0
+ */
+public class ValueCodingQosDefinitionParser implements BeanDefinitionParser {
+
+    @SuppressWarnings("DuplicatedCode")
+    @Override
+    public BeanDefinition parse(Element element, @Nonnull ParserContext parserContext) {
+        String serviceName = (String) BeanDefinitionParserUtil.mayResolveSpel(
+                parserContext, element.getAttribute("service-name")
+        );
+        String qosHandlerName = (String) BeanDefinitionParserUtil.mayResolveSpel(
+                parserContext, element.getAttribute("qos-handler-name")
+        );
+        String handlerRef = (String) BeanDefinitionParserUtil.mayResolveSpel(
+                parserContext, element.getAttribute("handler-ref")
+        );
+        String semRef = (String) BeanDefinitionParserUtil.mayResolveSpel(
+                parserContext, element.getAttribute("sem-ref")
+        );
+
+        BeanDefinitionParserUtil.makeSureBeanNameNotDuplicated(parserContext, qosHandlerName);
+        BeanDefinitionParserUtil.makeSureBeanNameNotDuplicated(parserContext, serviceName);
+
+        ManagedMap<String, BeanReference> valueCodingHandlerMap = new ManagedMap<>();
+        valueCodingHandlerMap.put(handlerRef, new RuntimeBeanReference(handlerRef));
+
+        BeanDefinitionBuilder valueCodingQosHandlerBuilder = BeanDefinitionBuilder.rootBeanDefinition(
+                ValueCodingQosHandlerImpl.class
+        );
+        valueCodingQosHandlerBuilder.getRawBeanDefinition().setAutowireMode(AbstractBeanDefinition.AUTOWIRE_CONSTRUCTOR);
+        ConstructorArgumentValues valueCodingQosHandlerConstructorArgumentValues = new ConstructorArgumentValues();
+        valueCodingQosHandlerConstructorArgumentValues.addIndexedArgumentValue(0, valueCodingHandlerMap);
+        valueCodingQosHandlerBuilder.getRawBeanDefinition().setConstructorArgumentValues(
+                valueCodingQosHandlerConstructorArgumentValues
+        );
+        valueCodingQosHandlerBuilder.setScope(BeanDefinition.SCOPE_SINGLETON);
+        valueCodingQosHandlerBuilder.setLazyInit(false);
+        parserContext.getRegistry().registerBeanDefinition(
+                qosHandlerName, valueCodingQosHandlerBuilder.getBeanDefinition()
+        );
+
+        BeanDefinitionBuilder valueCodingQosServiceBuilder = BeanDefinitionBuilder.rootBeanDefinition(
+                ValueCodingQosServiceImpl.class
+        );
+        valueCodingQosServiceBuilder.getRawBeanDefinition().setAutowireMode(AbstractBeanDefinition.AUTOWIRE_CONSTRUCTOR);
+        ConstructorArgumentValues valueCodingQosServiceConstructorArgumentValues = new ConstructorArgumentValues();
+        valueCodingQosServiceConstructorArgumentValues.addIndexedArgumentValue(
+                0, new RuntimeBeanReference(qosHandlerName)
+        );
+        valueCodingQosServiceConstructorArgumentValues.addIndexedArgumentValue(
+                1, new RuntimeBeanReference(semRef)
+        );
+        valueCodingQosServiceBuilder.getRawBeanDefinition().setConstructorArgumentValues(
+                valueCodingQosServiceConstructorArgumentValues
+        );
+        valueCodingQosServiceBuilder.setScope(BeanDefinition.SCOPE_SINGLETON);
+        valueCodingQosServiceBuilder.setLazyInit(false);
+        parserContext.getRegistry().registerBeanDefinition(serviceName, valueCodingQosServiceBuilder.getBeanDefinition());
+
+        return null;
+    }
+}